[VR-49] Fix: member delete api
- 교차지점이 없을 때 삭제가 되지 않는 이슈 해결

+++ src/main/java/com/vroomvroom/safemobis/service/MemberService.java
@@ -160,7 +160,9 @@ public class MemberService {
     public void delete() {
         Member member = findByUsername(getCurrentUsername());
         List<Intersection> intersections = getIntersections(member);
+        if (intersections.size() > 0) {
+            intersectionRepository.deleteAll(intersections);
+        }
         memberRepository.delete(member);
-        intersectionRepository.deleteAll(intersections);
     }
 }
